Android: moves net/ to use long for JNI.
BUG=

Review URL: https://codereview.chromium.org/67133002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@234285 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/android/java/src/org/chromium/net/ProxyChangeListener.java
@@ -25,7 +25,7 @@ public class ProxyChangeListener {
     private static final String TAG = "ProxyChangeListener";
     private static boolean sEnabled = true;
 
-    private int mNativePtr;
+    private long mNativePtr;
     private Context mContext;
     private ProxyReceiver mProxyReceiver;
     private Delegate mDelegate;
@@ -66,7 +66,7 @@ public class ProxyChangeListener {
     }
 
     @CalledByNative
-    public void start(int nativePtr) {
+    public void start(long nativePtr) {
         assert mNativePtr == 0;
         mNativePtr = nativePtr;
         registerReceiver();
@@ -165,9 +165,9 @@ public class ProxyChangeListener {
      * See net/proxy/proxy_config_service_android.cc
      */
     @NativeClassQualifiedName("ProxyConfigServiceAndroid::JNIDelegate")
-    private native void nativeProxySettingsChangedTo(int nativePtr,
+    private native void nativeProxySettingsChangedTo(long nativePtr,
                                                      String host,
                                                      int port);
     @NativeClassQualifiedName("ProxyConfigServiceAndroid::JNIDelegate")
-    private native void nativeProxySettingsChanged(int nativePtr);
+    private native void nativeProxySettingsChanged(long nativePtr);
 }
